# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a API documentation for OB server.  You can find out more about Ocean Builders at [https://ocean.builders/](https://ocean.builders/)",
    "version": "0.7",
    "title": "Ocean Builder",
    "contact": {
      "email": "diego@scissorboy.com"
    }
  },
  "host": "oceanbuilders.herokuapp.com",
  "basePath": "/v1/api/",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/permissions/{id}": {
      "post": {
        "tags": [
          "Permission"
        ],
        "summary": "Add new permission",
        "description": "",
        "operationId": "addNewPermission",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Permission Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Name": {
                  "type": "string"
                },
                "Sets": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "Name": {
                        "type": "string"
                      },
                      "Permissions": {
                        "type": "object",
                        "properties": {
                          "Status": {
                            "type": "string",
                            "enum": [
                              "ON",
                              "OFF",
                              "EMERGENCY"
                            ]
                          },
                          "Name": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "example": {
                "Name": "close friend",
                "Sets": [
                  {
                    "Name": "MainControls",
                    "Permissions": [
                      {
                        "Status": "ON",
                        "Name": "changeOrientation"
                      },
                      {
                        "Status": "ON",
                        "Name": "raiseLowerStairs"
                      },
                      {
                        "Status": "ON",
                        "Name": "frostWindows"
                      },
                      {
                        "Status": "ON",
                        "Name": "modifyACTemp"
                      },
                      {
                        "Status": "ON",
                        "Name": "doorsControl"
                      },
                      {
                        "Status": "ON",
                        "Name": "windowsControl"
                      },
                      {
                        "Status": "ON",
                        "Name": "smartWindowsControl"
                      },
                      {
                        "Status": "ON",
                        "Name": "powerSupply"
                      },
                      {
                        "Status": "ON",
                        "Name": "waterSupply"
                      },
                      {
                        "Status": "ON",
                        "Name": "audioSystem"
                      },
                      {
                        "Status": "ON",
                        "Name": "stove"
                      },
                      {
                        "Status": "ON",
                        "Name": "oven"
                      }
                    ]
                  },
                  {
                    "Name": "Cameras",
                    "Permissions": [
                      {
                        "Status": "ON",
                        "Name": "kitchen"
                      },
                      {
                        "Status": "ON",
                        "Name": "livingRoom"
                      },
                      {
                        "Status": "ON",
                        "Name": "Bedroom"
                      }
                    ]
                  },
                  {
                    "Name": "Lighting",
                    "Permissions": [
                      {
                        "Status": "ON",
                        "Name": "turnLights"
                      },
                      {
                        "Status": "ON",
                        "Name": "modifyIntensity"
                      },
                      {
                        "Status": "ON",
                        "Name": "switchLightscene"
                      },
                      {
                        "Status": "ON",
                        "Name": "addPersonalLightscene"
                      },
                      {
                        "Status": "ON",
                        "Name": "addGlobalLightscene"
                      },
                      {
                        "Status": "ON",
                        "Name": "manageGlobalLightscene"
                      }
                    ]
                  },
                  {
                    "Name": "AccessManagement",
                    "Permissions": [
                      {
                        "Status": "ON",
                        "Name": "inviteGuests"
                      }
                    ]
                  },
                  {
                    "Name": "SeapodInfo",
                    "Permissions": [
                      {
                        "Status": "ON",
                        "Name": "serialNumber"
                      },
                      {
                        "Status": "ON",
                        "Name": "modelInfoAndSpec"
                      }
                    ]
                  }
                ]
              }
            }
          },
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "permission data is required, Invalid seapod Id, Duplicated Permission Name"
          },
          "403": {
            "description": "Access denied. Not the Owner of the seapod"
          },
          "404": {
            "description": "Permission Not Found, Seapod Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      },
      "put": {
        "tags": [
          "Permission"
        ],
        "summary": "Update permission",
        "description": "",
        "operationId": "updatePermission",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Permission Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "isDefault": {
                  "type": "boolean"
                },
                "Name": {
                  "type": "string"
                },
                "Sets": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "Name": {
                        "type": "string"
                      },
                      "Permissions": {
                        "type": "object",
                        "properties": {
                          "Status": {
                            "type": "string",
                            "enum": [
                              "ON",
                              "OFF",
                              "EMERGENCY"
                            ]
                          },
                          "Name": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "example": {
                "isDefault": false,
                "Name": "close friend",
                "Sets": [
                  {
                    "Name": "MainControls",
                    "Permissions": [
                      {
                        "Status": "OFF",
                        "Name": "changeOrientation"
                      },
                      {
                        "Status": "OFF",
                        "Name": "raiseLowerStairs"
                      },
                      {
                        "Status": "OFF",
                        "Name": "frostWindows"
                      },
                      {
                        "Status": "ON",
                        "Name": "modifyACTemp"
                      },
                      {
                        "Status": "ON",
                        "Name": "doorsControl"
                      },
                      {
                        "Status": "ON",
                        "Name": "windowsControl"
                      },
                      {
                        "Status": "ON",
                        "Name": "smartWindowsControl"
                      },
                      {
                        "Status": "ON",
                        "Name": "powerSupply"
                      },
                      {
                        "Status": "ON",
                        "Name": "waterSupply"
                      },
                      {
                        "Status": "ON",
                        "Name": "audioSystem"
                      },
                      {
                        "Status": "ON",
                        "Name": "stove"
                      },
                      {
                        "Status": "ON",
                        "Name": "oven"
                      }
                    ]
                  },
                  {
                    "Name": "Cameras",
                    "Permissions": [
                      {
                        "Status": "ON",
                        "Name": "kitchen"
                      },
                      {
                        "Status": "ON",
                        "Name": "livingRoom"
                      },
                      {
                        "Status": "ON",
                        "Name": "Bedroom"
                      }
                    ]
                  },
                  {
                    "Name": "Lighting",
                    "Permissions": [
                      {
                        "Status": "ON",
                        "Name": "turnLights"
                      },
                      {
                        "Status": "ON",
                        "Name": "modifyIntensity"
                      },
                      {
                        "Status": "ON",
                        "Name": "switchLightscene"
                      },
                      {
                        "Status": "ON",
                        "Name": "addPersonalLightscene"
                      },
                      {
                        "Status": "ON",
                        "Name": "addGlobalLightscene"
                      },
                      {
                        "Status": "ON",
                        "Name": "manageGlobalLightscene"
                      }
                    ]
                  },
                  {
                    "Name": "AccessManagement",
                    "Permissions": [
                      {
                        "Status": "ON",
                        "Name": "inviteGuests"
                      }
                    ]
                  },
                  {
                    "Name": "SeapodInfo",
                    "Permissions": [
                      {
                        "Status": "ON",
                        "Name": "serialNumber"
                      },
                      {
                        "Status": "ON",
                        "Name": "modelInfoAndSpec"
                      }
                    ]
                  }
                ]
              }
            }
          },
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "permission data is required, Invalid seapod Id, Duplicated Permission Name, Default Permission cannot be updated"
          },
          "403": {
            "description": "Access denied. Not the Owner of the seapod"
          },
          "404": {
            "description": "Permission Not Found, Seapod Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      },
      "get": {
        "tags": [
          "Permission"
        ],
        "summary": "Get Permission By Id",
        "description": "",
        "operationId": "getPermissionById",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the permission"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid permission Id"
          },
          "404": {
            "description": "Permission Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/permissions/{permissionId}/seapod/{seapodId}/name": {
      "put": {
        "tags": [
          "Permission"
        ],
        "summary": "Rename permission",
        "description": "",
        "operationId": "renamePermission",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Permission Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Name": {
                  "type": "string"
                }
              },
              "example": {
                "Name": "close friend"
              }
            }
          },
          {
            "in": "path",
            "name": "permissionId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the permission"
          },
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "permission name is required, Invalid seapod Id, Invalid permission Id, Duplicated Permission Name, Default Permission cannot be renamed"
          },
          "403": {
            "description": "Access denied. Not the Owner of the seapod"
          },
          "404": {
            "description": "Permission Not Found, Seapod Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/permissions/{permissionId}/seapod/{seapodId}": {
      "delete": {
        "tags": [
          "Permission"
        ],
        "summary": "Delete permission",
        "description": "",
        "operationId": "deletePermission",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "permissionId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the permission"
          },
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid seapod Id, Invalid permission Id, Default Permission cannot be deleted"
          },
          "403": {
            "description": "Access denied. Not the Owner of the seapod"
          },
          "404": {
            "description": "Permission Not Found, Seapod Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/permissions/{seapodId}/user": {
      "put": {
        "tags": [
          "Permission"
        ],
        "summary": "Add permission to user",
        "description": "",
        "operationId": "addPermissionToUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User Id and Seapod Id",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "permissionId": {
                  "type": "string"
                }
              },
              "example": {
                "userId": "5ee8d9425d63472dd45d88d4",
                "permissionId": "5ee8ded5385e1722008f3366"
              }
            }
          },
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid seapod Id, Invalid permission Id, Invalid user Id"
          },
          "403": {
            "description": "Access denied. Not the Owner of the seapod"
          },
          "404": {
            "description": "Permission Not Found, Seapod Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      },
      "delete": {
        "tags": [
          "Permission"
        ],
        "summary": "Revoke permission from user",
        "description": "",
        "operationId": "revokePermissionFromUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User Id",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                }
              },
              "example": {
                "userId": "5ee8d9425d63472dd45d88d4"
              }
            }
          },
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid seapod Id, Invalid permission Id, Invalid user Id"
          },
          "403": {
            "description": "Access denied. Not the Owner of the seapod"
          },
          "404": {
            "description": "Permission Not Found, Seapod Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/permissions/{seapodId}/seapod": {
      "get": {
        "tags": [
          "Permission"
        ],
        "summary": "Get Seapod Permissions",
        "description": "",
        "operationId": "getSeapodPermissions",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid seapod Id"
          },
          "404": {
            "description": "Seapod Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
...
  }
}